# [hail] Reorganize `import_bgen` staged code; use lookup table to parse