This doesn't actually hugely improve performance:

Master:
```
2019-09-12 16:25:06,282: INFO: [1/1] Running import_bgen_force_count_all...
2019-09-12 16:26:26,427: INFO:     burn in: 80.14s
2019-09-12 16:27:46,816: INFO:     run 1: 80.39s
2019-09-12 16:29:10,637: INFO:     run 2: 83.82s
2019-09-12 16:30:33,742: INFO:     run 3: 83.11s
2019-09-12 16:31:53,968: INFO:     run 4: 80.22s
2019-09-12 16:33:18,855: INFO:     run 5: 84.89s
```

PR:
```
2019-09-12 16:46:20,424: INFO: [1/1] Running import_bgen_force_count_all...
2019-09-12 16:47:39,550: INFO:     burn in: 79.12s
2019-09-12 16:48:57,259: INFO:     run 1: 77.71s
2019-09-12 16:50:15,457: INFO:     run 2: 78.20s
2019-09-12 16:51:32,472: INFO:     run 3: 77.01s
2019-09-12 16:52:49,160: INFO:     run 4: 76.68s
2019-09-12 16:54:05,504: INFO:     run 5: 76.34s
```

However, it does make the generated much easier to profile, and
would make much more of difference if we support BGEN 1.3 with ZSTD.